Moved student run info and add student to run endpoints from RunAPIController to StudentAPIController.
